Only set 0 durations on the last release of a chain
Previously we were setting any release with a zero duration to end
today. But if two releases consecutive releases occurred at the
exact same time, the difference (duration) is 0. Which caused the
first of the releases to end today.

The check was specifically to allow current GA or beta or RC to have
a non zero duration so users could see that release still has some
time allocated to it, but as there's no further releases yet, it
would have a duration of 0.

Now, we'll just check the latest release on each chain, we're still
doing more than what we need to do (we check every release), but it's
less of a problem now.