Tags regexp to use comma separated and fix + position
The docs already suggested to use a comma to separate the tag from
date, and it's easier on the CLI to do comma than adding a tab but
also escaping the other git format syntax.

Also, plus was in the wrong location, so only minor version 0-9 were
allowed, not 0-99 etc.